# openapi.yaml
openapi: 3.0.3
info:
  title: Work2Work API
  version: 0.0.1
  description: OpenAPI specification for the current scheduling endpoints and response shapes.
servers:
  - url: http://localhost:8080
paths:
  /api/scheduling/shifts:
    post:
      summary: Create a shift
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateShiftRequest'
      responses:
        '201':
          description: Shift created
  /api/scheduling/shifts/{shiftId}:
    get:
      summary: Get a shift by id
      parameters:
        - $ref: '#/components/parameters/ShiftId'
      responses:
        '200':
          description: Shift details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShiftResponse'
    put:
      summary: Update a shift
      parameters:
        - $ref: '#/components/parameters/ShiftId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateShiftRequest'
      responses:
        '200':
          description: Updated shift
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShiftEntity'
    delete:
      summary: Soft-delete a shift
      parameters:
        - $ref: '#/components/parameters/ShiftId'
      responses:
        '200':
          description: Shift deleted
  /api/scheduling/employees:
    get:
      deprecated: true
      summary: Get grouped shifts in range
      parameters:
        - in: query
          name: companyId
          required: true
          schema:
            type: integer
        - in: query
          name: startDate
          required: true
          schema:
            type: string
            format: date
        - in: query
          name: endDate
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Employees with grouped daily shift buckets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EmployeeWithShifts'
  /api/scheduling/shifts/employees:
    get:
      summary: Get grouped shifts in range
      parameters:
        - in: query
          name: companyId
          required: true
          schema:
            type: integer
        - in: query
          name: startDate
          required: true
          schema:
            type: string
            format: date
        - in: query
          name: endDate
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Employees with grouped daily shift buckets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EmployeeWithShifts'
  /api/scheduling/shifts/date-position:
    get:
      summary: Get shifts grouped by date and position
      parameters:
        - in: query
          name: companyId
          required: true
          schema:
            type: integer
        - in: query
          name: startDate
          required: true
          schema:
            type: string
            format: date
        - in: query
          name: endDate
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Date buckets grouped by position, including empty buckets for company positions without shifts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DayPositionBucket'
  /api/scheduling/shifts/grouped:
    get:
      summary: Get shifts grouped by date and the requested grouping strategy
      parameters:
        - in: query
          name: companyId
          required: true
          schema:
            type: integer
        - in: query
          name: grouping
          required: true
          schema:
            type: string
            enum:
              - position_shift_timings
              - shift_timings
              - category_shift_timings
              - cat_shift_timings
        - in: query
          name: startDate
          required: true
          schema:
            type: string
            format: date
        - in: query
          name: endDate
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Grouped shifts in one normalized structure. Results are always grouped by date first, then by the requested grouping strategy under `shiftGroups`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupedShiftsResponse'
  /api/preferences/day:
    get:
      summary: Get day preference
      parameters:
        - in: query
          name: employeeId
          required: true
          schema:
            type: integer
        - in: query
          name: date
          required: true
          schema:
            type: string
            format: date
        - in: query
          name: companyId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DayPreferenceResponse'
    post:
      summary: Create or update day preference
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DayPreferenceRequest'
      responses:
        '200':
          description: OK
  /api/preferences/day/range:
    get:
      summary: Get day preferences in range
      parameters:
        - in: query
          name: employeeId
          required: true
          schema:
            type: integer
        - in: query
          name: startDate
          required: true
          schema:
            type: string
            format: date
        - in: query
          name: endDate
          required: true
          schema:
            type: string
            format: date
        - in: query
          name: companyId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DayPreferenceResponse'
  /api/preferences/day/repeat:
    post:
      summary: Create repeating day preference
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DayPreferenceRepeatRequest'
      responses:
        '200':
          description: OK
  /api/preferences/day/list:
    post:
      summary: Update multiple day preferences in a single call
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DayPreferenceListRequest'
      responses:
        '200':
          description: OK
  /api/preferences/resolved:
    get:
      summary: Get resolved preferences for a date range
      parameters:
        - in: query
          name: employeeId
          required: true
          schema:
            type: integer
        - in: query
          name: startDate
          required: true
          schema:
            type: string
            format: date
        - in: query
          name: endDate
          required: true
          schema:
            type: string
            format: date
        - in: query
          name: companyId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResolvedPreferenceResponse'
  /api/positions:
    get:
      summary: Get positions for a company
      parameters:
        - in: query
          name: companyId
          required: true
          schema:
            type: integer
        - in: query
          name: status
          required: false
          schema:
            type: string
            enum: [all, active, inactive]
            default: all
      responses:
        '200':
          description: A list of positions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PositionsResponse'
    post:
      summary: Create a new position
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePositionRequest'
      responses:
        '204':
          description: Position created successfully
  /api/positions/{id}:
    get:
      summary: Get an active position by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: query
          name: companyId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Position details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PositionSummary'
        '404':
          description: Position not found
    put:
      summary: Update an existing position
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: query
          name: companyId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePositionRequest'
      responses:
        '204':
          description: Position updated successfully
        '404':
          description: Position not found
    delete:
      summary: Soft-delete a position
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: query
          name: companyId
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Position soft-deleted successfully
        '404':
          description: Position not found
  /api/position-groups:
    get:
      summary: Get position groups for a company
      parameters:
        - in: query
          name: companyId
          required: true
          schema:
            type: integer
        - in: query
          name: status
          required: false
          schema:
            type: string
            enum: [all, active, inactive]
            default: all
      responses:
        '200':
          description: A list of position groups
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PositionGroupsResponse'
    post:
      summary: Create a new position group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePositionGroupRequest'
      responses:
        '204':
          description: Position group created successfully
        '400':
          description: One or more position ids were invalid for the company
  /api/position-groups/active:
    get:
      summary: Get all active position groups for a company
      parameters:
        - in: query
          name: companyId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: A list of active position groups
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PositionGroupsResponse'
  /api/position-groups/non-active:
    get:
      summary: Get all inactive position groups for a company
      parameters:
        - in: query
          name: companyId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: A list of inactive position groups
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PositionGroupsResponse'
  /api/position-groups/{id}:
    get:
      summary: Get an active position group by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: query
          name: companyId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Position group details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PositionGroupSummary'
        '404':
          description: Position group not found
    put:
      summary: Update an existing position group
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: query
          name: companyId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePositionGroupRequest'
      responses:
        '204':
          description: Position group updated successfully
        '400':
          description: One or more position ids were invalid for the company
        '404':
          description: Position group not found
    delete:
      summary: Soft-delete a position group
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: query
          name: companyId
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Position group soft-deleted successfully
        '404':
          description: Position group not found
components:
  parameters:
    ShiftId:
      in: path
      name: shiftId
      required: true
      schema:
        type: integer
  schemas:
    CreateShiftRequest:
      type: object
      required: [date, startTime, endTime, position]
      properties:
        employeeId:
          type: integer
        companyId:
          type: integer
        description:
          type: string
        date:
          type: string
          format: date
        startTime:
          type: string
          example: 9:00AM
        endTime:
          type: string
          example: 5:00PM
        duration:
          type: number
          format: float
        position:
          type: integer
        category:
          type: integer
        color:
          type: string
    UpdateShiftRequest:
      type: object
      required: [date, startTime, endTime, position]
      properties:
        employeeId:
          type: integer
        description:
          type: string
        date:
          type: string
          format: date
        startTime:
          type: string
          example: 10:00AM
        endTime:
          type: string
          example: 6:00PM
        duration:
          type: number
          format: float
        position:
          type: integer
        category:
          type: integer
        color:
          type: string
    ShiftResponse:
      type: object
      properties:
        shiftId:
          type: integer
        employeeId:
          type: integer
        companyId:
          type: integer
        description:
          type: string
        date:
          type: string
          format: date
        startTime:
          type: string
        endTime:
          type: string
        duration:
          type: number
          format: float
        isOvernight:
          type: boolean
        position:
          type: string
        category:
          type: string
        color:
          type: string
    ShiftEntity:
      type: object
      properties:
        shiftId:
          type: integer
        employeeId:
          type: integer
        scheduleId:
          type: integer
        companyId:
          type: integer
        description:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        duration:
          type: number
          format: float
        isOvernight:
          type: boolean
        requiredPositionId:
          type: integer
        categoryId:
          type: integer
        color:
          type: string
        isDeleted:
          type: boolean
        changedBy:
          type: integer
    Position:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
    GroupedShift:
      type: object
      properties:
        shiftId:
          type: integer
        startTime:
          type: string
        endTime:
          type: string
        position:
          type: string
        category:
          type: string
        description:
          type: string
        duration:
          type: number
          format: float
        color:
          type: string
    DayShiftBucket:
      type: object
      properties:
        date:
          type: string
          format: date
        shifts:
          type: array
          items:
            $ref: '#/components/schemas/GroupedShift'
    EmployeeWithShifts:
      type: object
      properties:
        employeeId:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        phones:
          type: array
          items:
            type: string
        availablePositions:
          type: array
          items:
            $ref: '#/components/schemas/Position'
        weeklyShifts:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/DayShiftBucket'
        totalHours:
          type: number
          format: float
        shiftCount:
          type: integer
    EmployeeScheduledShift:
      type: object
      properties:
        shiftId:
          type: integer
        employeeId:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        phones:
          type: array
          items:
            type: string
        startTime:
          type: string
        endTime:
          type: string
        category:
          type: string
        description:
          type: string
        duration:
          type: number
          format: float
        color:
          type: string
    PositionShiftBucket:
      type: object
      properties:
        position:
          type: string
          nullable: true
        shifts:
          type: array
          items:
            $ref: '#/components/schemas/EmployeeScheduledShift'
        shiftCount:
          type: integer
        totalDuration:
          type: number
          format: float
    PositionTimingBucket:
      type: object
      properties:
        position:
          type: string
          nullable: true
        shiftTimings:
          type: array
          items:
            $ref: '#/components/schemas/ShiftTimingBucket'
        shiftCount:
          type: integer
        totalDuration:
          type: number
          format: float
    ShiftTimingBucket:
      type: object
      properties:
        startTime:
          type: string
        endTime:
          type: string
        shifts:
          type: array
          items:
            $ref: '#/components/schemas/EmployeeScheduledShift'
        shiftCount:
          type: integer
        totalDuration:
          type: number
          format: float
    CategoryTimingBucket:
      type: object
      properties:
        category:
          type: string
          nullable: true
        shiftTimings:
          type: array
          items:
            $ref: '#/components/schemas/ShiftTimingGroup'
    ShiftTimingGroup:
      type: object
      properties:
        label:
          type: string
        shifts:
          type: array
          items:
            $ref: '#/components/schemas/EmployeeScheduledShift'
    DayShiftTimingBucket:
      type: object
      properties:
        date:
          type: string
          format: date
        shiftTimings:
          type: array
          items:
            $ref: '#/components/schemas/ShiftTimingGroup'
    DayPositionBucket:
      type: object
      properties:
        date:
          type: string
          format: date
        positions:
          type: array
          items:
            $ref: '#/components/schemas/PositionShiftBucket'
        shiftCount:
          type: integer
        totalDuration:
          type: number
          format: float
    GroupedShiftsResponse:
      type: object
      properties:
        dates:
          type: array
          items:
            $ref: '#/components/schemas/GroupedShiftDate'
    GroupedShiftDate:
      type: object
      properties:
        date:
          type: string
          format: date
        shiftGroups:
          type: array
          items:
            $ref: '#/components/schemas/ShiftGroup'
    ShiftGroup:
      type: object
      properties:
        label:
          type: string
          nullable: true
        shiftGroups:
          type: array
          items:
            $ref: '#/components/schemas/ShiftGroup'
        shifts:
          type: array
          items:
            $ref: '#/components/schemas/EmployeeScheduledShift'
    DayPreferenceRequest:
      type: object
      required: [employeeId, companyId, date]
      properties:
        employeeId:
          type: integer
        companyId:
          type: integer
        date:
          type: string
          format: date
        prefs:
          type: string
        compression:
          type: integer
        editedBy:
          type: integer
        isDayPrefs:
          type: boolean
    DayPreferenceResponse:
      type: object
      properties:
        date:
          type: string
          format: date
        prefs:
          type: string

        day:
          type: string
        isDayPrefs:
          type: boolean
    DayPreferenceRepeatRequest:
      type: object
      required: [employeeId, companyId, date, repeatCount]
      properties:
        employeeId:
          type: integer
        companyId:
          type: integer
        date:
          type: string
          format: date
        prefs:
          type: string
        compression:
          type: integer
        editedBy:
          type: integer
        repeatCount:
          type: integer
        isDayPrefs:
          type: boolean
    DayPreferenceListRequest:
      type: object
      required: [preferences]
      properties:
        preferences:
          type: array
          items:
            $ref: '#/components/schemas/DayPreferenceRequest'
    ResolvedPreferenceResponse:
      type: object
      properties:
        date:
          type: string
          format: date
        prefs:
          type: string
        preferenceType:
          type: string
        day:
          type: string
    PositionSummary:
      type: object
      properties:
        positionId:
          type: integer
        description:
          type: string
    CreatePositionRequest:
      type: object
      required:
        - companyId
        - description
      properties:
        companyId:
          type: integer
        description:
          type: string
    UpdatePositionRequest:
      type: object
      required:
        - description
      properties:
        description:
          type: string
    PositionsResponse:
      type: object
      properties:
        positions:
          type: array
          items:
            $ref: '#/components/schemas/PositionSummary'
    CreatePositionGroupRequest:
      type: object
      required:
        - companyId
        - description
        - positionIds
      properties:
        companyId:
          type: integer
        description:
          type: string
        positionIds:
          type: array
          items:
            type: integer
    UpdatePositionGroupRequest:
      type: object
      required:
        - description
        - positionIds
      properties:
        description:
          type: string
        positionIds:
          type: array
          items:
            type: integer
    PositionGroupSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        positions:
          type: array
          items:
            $ref: '#/components/schemas/PositionSummary'
    PositionGroupsResponse:
      type: object
      properties:
        positionGroups:
          type: array
          items:
            $ref: '#/components/schemas/PositionGroupSummary'